Disable "sites" and Enable "djangotoolbox"
By default, the application ``django.contrib.sites`` won't behave
well with the Django MongoDB engine. Under the hood, it boils down
to differences in primary keys, which are strings with MongoDB,
and integers elsewhere. It would of course be more elegant to fix
``sites`` in the first place, but for the sake of simplicity, we
will just disable it since we don't need it for simple apps.

Also, we need ``djangotoolbox`` to make user editing in the
admin site work correctly. Long story short, ``djangotoolbox``
allows us to do some JOINs on non-relational databases.